rag/requirements.txt
openai>=1.12.0
pypdf>=4.0.0
chromadb>=0.4.22
langchain-text-splitters>=0.0.1
requests>=2.31.0
python-dotenv>=1.0.0
